+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All Notable changes to `bakame/html-table` will be documented in this file.
 
+## [Next](https://github.com/bakame-php/http-structured-fields/compare/0.1.0...master) - TBD
+
+### Added
+
+- `Parser::withFormatter`
+- `Parser::withoutFormatter`
+
+### Fixed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- None
+
 ## [0.1.0] - 2023-09-22
 
 **Initial release!**